third_party/blink/web_tests/external/wpt/encoding-detection/ja-EUC-JP.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Check detection result
  assert_equals: Expected EUC-JP expected "EUC-JP" but got "GBK"
Harness: the test ran to completion.
